chore: move learn to team workspace #453
