+++ src/app/(routes)/studio/(teams)/[team_slug]/(workspace)/layout.tsx
