+++ src/app/(routes)/studio/(teams)/[team_slug]/(workspace)/oops/page.tsx
