+++ src/app/(routes)/studio/(teams)/[team_slug]/(workspace)/workspace/(home)/layout.tsx
