+++ src/app/(routes)/studio/(teams)/[team_slug]/(workspace)/workspace/(learn)/learn/(playbook)/layout.tsx
@@ -3,21 +3,14 @@ import { redirect } from 'next/navigation';
 import { authOptions } from '#/lib/auth';
 import { getTeams } from '#/lib/team/get-teams';
 
-import { NavBar } from '#/components/layout/navbar';
-import { SiteFooter } from '#/components/layout/site-footer';
-
 interface PlaybookRootLayoutProps {
   children: React.ReactNode;
-  params: {
-    team_slug: string;
-  };
 }
 
 export default async function PlaybookRootLayout({
   children,
-  params,
 }: PlaybookRootLayoutProps) {
-  const { user, teams } = await getTeams();
+  const { user } = await getTeams();
 
   if (!user) {
     redirect(authOptions?.pages?.signIn || '/login');
@@ -25,9 +18,7 @@ export default async function PlaybookRootLayout({
 
   return (
     <div className="flex min-h-screen flex-col">
-      <NavBar user={user} teams={teams} activeTeamSlug={params.team_slug} />
       <div className="container flex-1">{children}</div>
-      <SiteFooter />
     </div>
   );
 }
