+++ src/app/(routes)/studio/(teams)/[team_slug]/(workspace)/workspace/(learn)/learn/(playbook)/loading.tsx
