+++ src/app/(routes)/studio/(teams)/[team_slug]/(workspace)/workspace/(learn)/learn/(playbooks)/layout.tsx
@@ -7,8 +7,6 @@ import { APP_BP } from '#/lib/const';
 import { getTeams } from '#/lib/team/get-teams';
 
 import { DashboardNav } from '#/components/layout/nav';
-import { NavBar } from '#/components/layout/navbar';
-import { SiteFooter } from '#/components/layout/site-footer';
 import { toast } from '#/components/ui/use-toast';
 
 interface PlaybooksLayoutProps {
@@ -43,13 +41,6 @@ export default async function PlaybooksLayout({
 
   return (
     <div className="flex min-h-screen flex-col space-y-6">
-      <NavBar
-        user={user}
-        teams={teams}
-        activeTeamSlug={params.team_slug}
-        scroll={false}
-      />
-
       <div className="container grid flex-1 gap-12 md:grid-cols-[200px_1fr]">
         <aside className="hidden w-[200px] flex-col py-8 md:flex">
           <DashboardNav items={config.sidebarNav} />
@@ -58,7 +49,6 @@ export default async function PlaybooksLayout({
           {children}
         </main>
       </div>
-      <SiteFooter />
     </div>
   );
 }
