+++ src/app/(routes)/studio/(teams)/[team_slug]/(workspace)/workspace/route.ts
@@ -1,0 +1,16 @@
+import { redirect } from 'next/navigation';
+import { NextRequest } from 'next/server';
+
+import { APP_BP } from '#/lib/const';
+
+interface Props {
+  params: {
+    team_slug: string;
+  };
+}
+
+export async function GET(request: NextRequest, props: Props) {
+  const { params } = props;
+
+  redirect(`${APP_BP}/${params.team_slug}/workspace/home`);
+}
