+++ src/app/(routes)/studio/(teams)/[team_slug]/route.ts
