+++ src/components/learn/courses/course-progress-button.tsx
@@ -46,7 +46,7 @@ export const CourseProgressButton = ({
 
       if (!isCompleted && nextChapterId) {
         router.push(
-          `${APP_BP}/learn/courses/${courseId}/chapters/${nextChapterId}`,
+          `${APP_BP}/workspace/learn/courses/${courseId}/chapters/${nextChapterId}`,
         );
       }
