+++ src/components/learn/courses/video-player.tsx
@@ -58,7 +58,7 @@ export const VideoPlayer = ({
 
         if (nextChapterId) {
           router.push(
-            `${APP_BP}/learn/courses/${courseId}/chapters/${nextChapterId}`,
+            `${APP_BP}/workspace/learn/courses/${courseId}/chapters/${nextChapterId}`,
           );
         }
       }
